NEW:Creacion de pagina con appBar menu y redirreccion a home y aboutUS
