+++ app/components/Header/Header.tsx
@@ -1,0 +1,16 @@
+import AppBar from "../AppBar/AppBar";
+
+
+
+
+
+const Header = ({children}) => {
+    return (
+        <>
+            <AppBar/>
+            {children}
+        </>
+    );
+}
+
+export default Header;
