+++ app/layout.js
@@ -1,17 +1,20 @@
-import { Inter } from 'next/font/google'
+
 import './globals.css'
 
-const inter = Inter({ subsets: ['latin'] })
+import Header from './components/Header/Header'
 
-export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
+// const inter = Inter({ subsets: ['latin'] })
+
+// export const metadata = {
+//   title: 'Create Next App',
+//   description: 'Generated by create next app',
+// }
 
 export default function RootLayout({ children }) {
   return (
-    <html lang="en">
-      <body className={inter.className}>{children}</body>
-    </html>
+    <div>
+      <Header/>
+      {children}
+    </div>
   )
 }
