+++ app/page.js
@@ -1,12 +1,17 @@
-import AppBarMenu from "./components/AppBar/AppBar"
-import WelcomeSection from "./components/Hero/WelcomeSection"
+import RootLayout from "./layout"
+import WelcomeSection from "./components/WelcomeSection/WelcomeSection"
+import StartContentHome from "./components/Home/StartContentHome"
 
 
 export default function Home() {
   return (
     <main>
       <div>
-        <WelcomeSection />
+          <RootLayout>
+            <WelcomeSection>
+              <StartContentHome />
+            </WelcomeSection>
+          </RootLayout>
       </div>
     </main>
   )
